+++ src/text.tsx
@@ -51,7 +51,10 @@ const useStyles = createUseStyles((theme) => {
         truncate: {
             '-webkit-line-clamp': lineClamp,
             lineClamp,
-            wordBreak: ({truncate}) => (truncate === 1 || truncate === true ? 'break-all' : 'normal'),
+            wordBreak: ({wordBreak}) => (wordBreak ? 'break-all' : 'normal'),
+            '@supports (overflow-wrap: anywhere)': {
+                wordBreak: ({truncate}) => (truncate === 1 || truncate === true ? 'break-all' : 'normal'),
+            },
             display: 'box',
             boxOrient: 'vertical',
             overflow: 'hidden',